requirements.txt
appdirs==1.4.4
astroid==3.0.3
black==24.1.1
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
fonttools==4.49.0
hidapi==0.14.0
isort==5.13.2
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
numpy
opencv-python==4.9.0.80
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pylint==3.0.3
pyparsing==3.1.1
pyshark==0.6
python-dateutil==2.8.2
python-pcapng==2.1.1
scapy==2.5.0
six==1.16.0
tabulate==0.9.0
termcolor==2.4.0
tomlkit==0.12.3
typing_extensions==4.9.0
